+++ src/app/containers/app-navbar/app-navbar-menu/app-navbar-menu.component.ts
@@ -21,8 +21,9 @@ enum Key {
       <i *ngIf="appVersion.isNewAvailable" class="pulse update-indicator fa fa-dot-circle-o text-success"></i>
     </button>
     <div class="panel panel-default menu-dropdown"
-      *ngIf="!hide">
-      <div class="menu-backdrop" (click)="hideMenu()"></div>
+      [class.slideInDown]="!hide"
+      >
+      <div class="menu-backdrop" *ngIf="!hide" (click)="hideMenu()"></div>
       <div class="list-group">
         <div *ngIf="appVersion.isNewAvailable" class="list-group-item">
           <button class="btn btn-success" title="click to update Echoes"
@@ -33,7 +34,7 @@ enum Key {
         <a class="list-group-item" href="http://github.com/orizens/echoes-player" target="_blank">
           <i class="fa fa-github"></i> Source Code @Github
         </a>
-        <a class="list-group-item" href="https://travis-ci.org/orizens/echoes-player" target="_blank">
+        <a class="list-group-item" *ngIf="!hide" href="https://travis-ci.org/orizens/echoes-player" target="_blank">
           <img src="https://travis-ci.org/orizens/echoes-player.svg?branch=master">
         </a>
         <div class="list-group-item" target="_blank">