+++ src/app/containers/youtube-videos/player-search.component.ts
@@ -18,13 +18,22 @@ import { Observable } from 'rxjs/Observable';
   selector: 'player-search',
   encapsulation: ViewEncapsulation.None,
   styleUrls: [ './player-search.scss' ],
+  // <input placeholder="Explore Media" id="media-search"
+  //         typeahead (typeaheadSelected)="handleSelectSuggestion($event)"
+  //         type="search" class="form-control" autocomplete="off"
+  //         [value]="query.query" #mediaSearch name="mediaSearch"
+  //         (input)="onQueryChange(mediaSearch.value)"
+  //         >
   template: `
     <form class="navbar-form form-search" id="media-explorer"
       #form="ngForm"
       (ngSubmit)="onSearch(mediaSearch.value)">
       <div class="form-group clearfix">
         <input placeholder="Explore Media" id="media-search"
-          typeahead (typeaheadSelected)="handleSelectSuggestion($event)"
+          ngxTypeahead
+          [taUrl]="'http://suggestqueries.google.com/complete/search'"
+          [taParams]="params"
+          (taSelected)="handleSelectSuggestion($event)"
           type="search" class="form-control" autocomplete="off"
           [value]="query.query" #mediaSearch name="mediaSearch"
           (input)="onQueryChange(mediaSearch.value)"
@@ -44,6 +53,13 @@ export class PlayerSearch {
   @Output() typing = new EventEmitter<string>();
 
   @ViewChild('mediaSearch') mediaSearch;
+
+  params = {
+    hl: 'en',
+    ds: 'yt',
+    xhr: 't',
+    client: 'youtube'
+  };
 
   constructor() { }