+++ src/app/core/components/now-playing/now-playlist/index.ts
@@ -1,1 +1,2 @@
 export * from './now-playlist.component';
+export * from './now-playlist-track.component';